BONUS: share `waitForPingContent` across the tests
