+++ samples/android/app/src/androidTest/java/org/mozilla/samples/gleancore/pings/DeletionRequestPingTest.kt
@@ -20,19 +20,10 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mozilla.samples.gleancore.MainActivity
-import org.mozilla.samples.gleancore.getPingServer
 import androidx.test.uiautomator.UiDevice
-import androidx.work.WorkInfo
-import androidx.work.WorkManager
-import androidx.work.testing.WorkManagerTestInitHelper
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.withTimeout
 import mozilla.telemetry.glean.testing.GleanTestLocalServer
-import org.json.JSONObject
 import org.junit.Assert.assertNotEquals
-import org.junit.Before
 import org.mozilla.samples.gleancore.getPingServerPort
-import java.util.concurrent.TimeUnit
 
 @RunWith(AndroidJUnit4::class)
 class DeletionRequestPingTest {
@@ -40,58 +31,10 @@ class DeletionRequestPingTest {
     val activityRule: ActivityTestRule<MainActivity> = ActivityTestRule(MainActivity::class.java)
 
     @get:Rule
-    val gleanRule = GleanTestLocalServer(getPingServerPort())
+    val gleanRule = GleanTestLocalServer(context, getPingServerPort())
 
     private val context: Context
         get() = ApplicationProvider.getApplicationContext()
-
-    @Before
-    fun clearWorkManager() {
-        WorkManagerTestInitHelper.initializeTestWorkManager(context)
-    }
-
-    @Suppress("ReturnCount")
-    private fun waitForPingContent(pingName: String): JSONObject? {
-        val server = getPingServer()
-
-        val request = server.takeRequest(20L, TimeUnit.SECONDS)
-        if (request == null) {
-            return null
-        }
-        val docType = request.path.split("/")[3]
-        if (pingName != docType) {
-            return null
-        }
-
-        return JSONObject(request.body.readUtf8())
-    }
-
-    /**
-     * Sadly, the WorkManager still requires us to manually trigger the upload job.
-     * This function goes through all the active jobs by Glean (there should only be
-     * one!) and triggers them.
-     */
-    private fun triggerEnqueuedUpload(tag: String) {
-        // The tag is really internal to PingUploadWorker, but we can't do much more
-        // than copy-paste unless we want to increase our API surface.
-        val reasonablyHighCITimeoutMs = 5000L
-
-        runBlocking {
-            withTimeout(reasonablyHighCITimeoutMs) {
-                do {
-                    val workInfoList = WorkManager.getInstance(context).getWorkInfosByTag(tag).get()
-                    workInfoList.forEach { workInfo ->
-                        if (workInfo.state === WorkInfo.State.ENQUEUED) {
-                            // Trigger WorkManager using TestDriver
-                            val testDriver = WorkManagerTestInitHelper.getTestDriver(context)
-                            testDriver?.setAllConstraintsMet(workInfo.id)
-                            return@withTimeout
-                        }
-                    }
-                } while (true)
-            }
-        }
-    }
 
     @Test
     fun validateDeletionRequestPing() {
@@ -105,21 +48,8 @@ class DeletionRequestPingTest {
             .perform(closeSoftKeyboard())
             .perform(click())
 
-        // Wait for the upload job to be present and trigger it.
-        Thread.sleep(1000) // FIXME: for some reason, without this, WorkManager won't find the job
-        triggerEnqueuedUpload("mozac_service_glean_deletion_ping_upload_worker")
-
         // We might receive previous baseline or events ping, let's ignore that
-        var retries = 3
-        var ping: JSONObject? = null
-        while (retries > 0) {
-            ping = waitForPingContent("deletion-request")
-            if (ping != null) {
-                break
-            }
-            retries -= 1
-        }
-        val deletionPing = ping!!
+        val deletionPing = waitForPingContent("deletion-request")!!
 
         // Validate the received data.
         assertEquals("deletion-request", deletionPing.getJSONObject("ping_info")["ping_type"])
@@ -133,10 +63,6 @@ class DeletionRequestPingTest {
 
         // Move it to background.
         device.pressHome()
-
-        // Wait for the upload job to be present and trigger it.
-        Thread.sleep(1000) // FIXME: for some reason, without this, WorkManager won't find the job
-        triggerEnqueuedUpload("mozac_service_glean_ping_upload_worker")
 
         // Validate the received data.
         val baselinePing = waitForPingContent("baseline")!!
